\name{cwhstat-package}
\alias{cwhstat-package}
\docType{package}
\title{
cwh's mathematical functions
}
\description{
Miscellaneous statistical functions for general use
}
\details{
\tabular{ll}{
Package: \tab cwhstat\cr
Version: \tab 4.2.0\cr
Date: \tab 2006-01-26\cr
Depends: \tab R (>= 2.0.0)\cr
BundleDescription: \tab Miscellaneous functions for general use\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.wsl.ch/staff/christian.hoffmann/\cr
}

Index:
\preformatted{
boxcox                  Box-Cox transformation and inverse
dinvgauss               Inverse Gaussian Distribution
dpoisgam                Poisson Gamma Distribution
f.log                   Determine an optimized offset s and return
                        log10(data+s).
jitterNA                Jitter entities with NA
my.table.NA             Tabulate data, with extra rows and columns.
neg.bin.gof             Approximate a Negative binomial distribution.
qnorm.ap16              Approximation to the inverse normal
                        distribution function.
qres.binom              Randomized quantile residuals
shapiro.wilk.test       Shapiro-Wilk Normality Test
smoothed.df             Fit cumulative distribution from kernel
                        estimate.
tweedie                 Tweedie Distributions
tweedief                Tweedie generalized linear model family
weighted.mean1          Weighted mean1
weighted.median         Weighted median
}
}
\author{
Christian W. Hoffmann <christian.hoffmann@wsl.ch>

Maintainer: Christian W. Hoffmann <christian.hoffmann@wsl.ch>
}
\keyword{ package }
