\name{invgauss}
\alias{dinvgauss}
\alias{pinvgauss}
\alias{qinvgauss}
\alias{rinvgauss}
\title{Inverse Gaussian Distribution}
\description{
  Density, cumulative probability, quantiles and random generation for the inverse Gaussian distribution.

	See full document: 	\url{http://www.maths.uq.edu.au/~gks/s/invgauss.html}
}
\author{Gordon Smyth, \email{gks@maths.uq.edu.au} }
\seealso{
	\url{http://www.maths.uq.edu.au/~gks/s/qres.html}
}
\keyword{distribution}
