\name{negbingof}
\alias{neg.bin.gof}
\title{Approximate a Negative binomial distribution.}
\description{
  \code{neg.bin.gof} determines negative binomial distribution parameter 'k' by successive approximation
}
\usage{
neg.bin.gof(data)
}
\arguments{
  \item{data}{vector of data.}
}
\value{
	Prints statistics of approximated distribution.
}
\author{?}
\keyword{distribution}
