\name{poisgam}
\alias{dpoisgam}
\alias{ppoisgam}
\alias{qpoisgam}
\alias{ppoiscc}
\alias{rpoisgam}
\title{Poisson Gamma Distribution}
\description{
  Density, cumulative probability function and random variates for the Poisson-gamma or
     compound Poisson distribution. The Poison-gamma distribution is a Tweedie distribution
     with index p between 1 and 2.  

	See full document: 	\url{http://www.maths.uq.edu.au/~gks/s/poisgam.html}
}
\author{Gordon Smyth, \email{gks@maths.uq.edu.au} }
\seealso{
 \url{http://www.maths.uq.edu.au/~gks/s/tweedie.html} Tweedie Distributions,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/tweedief.html}	Tweedie family,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/qres.html} Randomized quantile residuals,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/poisgam.html} Poison-gamma Distribution,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/invgauss.html} Inverse Gaussian Distribution.\newline
}
\keyword{distribution}
