\name{Randomized quantile residuals}
\alias{qres.binom}
\alias{qres.gamma}
\alias{qres.invgauss}
\alias{qres.nbinom}
\alias{qres.pois}
\alias{qres.tweedie}
\title{Randomized quantile residuals}
\description{
  Computes randomized quantile residuals for binomial, Poisson, negative
	binomial, gamma and inverse Gaussian generalized linear models.

	See full document: 	\url{http://www.maths.uq.edu.au/~gks/s/qres.html}
}
\author{Gordon Smyth, \email{gks@maths.uq.edu.au}}
\seealso{
 \url{http://www.maths.uq.edu.au/~gks/s/tweedie.html} Tweedie Distributions,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/tweedief.html}	Tweedie family,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/poisgam.html} Poison-gamma Distribution,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/invgauss.html} Inverse Gaussian Distribution.\newline
}
\keyword{distribution}
