\name{tweedie}
\alias{tweedie}  # why is this necessary?
\alias{dtweedie}
\alias{ptweedie}
\alias{qtweedie}
\alias{saddle.tweedie}
\alias{devi.tweedie}
\alias{tail.tweedie}
\title{Tweedie Distributions}
\description{
  Density, cumulative distribution function and quantiles for the
	Tweedie distributions. Includes the normal, Poison, Poison-gamma and inverse-Gaussian distributions as special cases. 

	See full document: 	\url{http://www.maths.uq.edu.au/~gks/s/tweedief.html}
}
\author{Gordon Smyth, \email{gks@maths.uq.edu.au} }
\seealso{
 \url{http://www.maths.uq.edu.au/~gks/s/tweedief.html} Tweedie family,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/qres.html} Randomized quantile residuals,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/poisgam.html} Poison-gamma Distribution,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/invgauss.html} Inverse Gaussian Distribution.\newline
}
\keyword{distribution}
