\name{tweedief}
\alias{tweedief}
\title{Tweedie generalized linear model family}
\description{
  Produces a generalized linear model family object with any power
	variance function and any power link. Includes the Gaussian, Poisson,
	gamma and inverse-Gaussian families as special cases.
		 
	See full document: 	\url{http://www.maths.uq.edu.au/~gks/s/tweedief.html}
}
\author{Gordon Smyth, \email{gks@maths.uq.edu.au} }
\seealso{
 \url{http://www.maths.uq.edu.au/~gks/s/tweedie.html} Tweedie Distributions,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/qres.html} Randomized quantile residuals,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/poisgam.html} Poison-gamma Distribution,\newline
 \url{http://www.maths.uq.edu.au/~gks/s/invgauss.html} Inverse Gaussian Distribution.\newline
}
\keyword{distribution}
